move to ingest_utils
