+++ app/src/util/ingest_utils.py
@@ -1,5 +1,6 @@
 import getopt
 import logging
+import re
 from logging import Logger
 from typing import Callable, Optional, Sequence
 
@@ -8,6 +9,8 @@ from sqlalchemy import delete, select
 from src.adapters import db
 from src.app_config import app_config
 from src.db.models.document import Chunk, Document
+
+logger = logging.getLogger(__name__)
 
 
 def _drop_existing_dataset(db_session: db.Session, dataset: str) -> bool:
@@ -96,3 +99,122 @@ def add_embeddings(
         assert (
             chunk.tokens <= embedding_model.max_seq_length
         ), f"Text too long for embedding model: {chunk.tokens} tokens: {len(chunk.content)} chars: {chunk.content[:80]}...{chunk.content[-50:]}"
+
+
+def deconstruct_list(
+    markdown: str, has_intro_sentence: bool = False, list_delimiter: str = r"^( *[\-\*\+] )"
+) -> tuple[str, list[str]]:
+    "Deconstruct a list of items into an introductory sentence and a list of items"
+    splits = re.split(list_delimiter, markdown, flags=re.MULTILINE)
+    if has_intro_sentence or not markdown.startswith(list_delimiter):
+        intro_sentence = splits[0]
+        splits = splits[1:]
+    else:
+        intro_sentence = ""
+
+    # Reconstruct each list item
+    list_items = [(splits[a] + splits[a + 1]) for a in range(0, len(splits), 2)]
+    return intro_sentence, list_items
+
+
+def reconstruct_list(
+    token_limit: int,
+    intro_sentence: str,
+    list_items: Sequence[str],
+    join_delimiter: str = "",
+) -> list[str]:
+    "Reconstruct a list of items into chunks (with same intro_sentence) having up to a character limit each"
+    # Before the set of list items, there should be a blank line
+    intro_sentence = _ensure_blank_line_suffix(intro_sentence)
+
+    item_max_seq_length = token_limit - len(tokenize(intro_sentence))
+    assert item_max_seq_length > 0, item_max_seq_length
+    chunks = [
+        intro_sentence + some_list_items
+        for some_list_items in _join_up_to_max_seq_length(
+            list_items, item_max_seq_length, delimiter=join_delimiter
+        )
+    ]
+    assert all(len(tokenize(chunk)) <= token_limit for chunk in chunks)
+    return chunks
+
+
+def _ensure_blank_line_suffix(text: str) -> str:
+    "Return text with a blank line after the text"
+    if not text.endswith("\n\n"):
+        if text.endswith("\n"):
+            text += "\n"
+        else:
+            text += "\n\n"
+    return text
+
+
+def deconstruct_table(
+    markdown: str, has_intro_sentence: bool = False, list_delimiter: str = r"^(\| )"
+) -> tuple[str, str, list[str]]:
+    "Deconstruct a table of items into an introductory sentence, a table header, and a list of items"
+    intro_sentence, list_items = deconstruct_list(markdown, has_intro_sentence, list_delimiter)
+
+    if list_items[1].startswith("| --- |"):
+        table_header = list_items.pop(0) + list_items.pop(0)
+    else:  # Assume no table header
+        table_header = ""
+
+    return intro_sentence, table_header, list_items
+
+
+def reconstruct_table(
+    token_limit: int,
+    intro_sentence: str,
+    table_header: str,
+    table_rows: Sequence[str],
+    join_delimiter: str = "",
+) -> list[str]:
+    "Reconstruct a table of items into chunks (with intro_sentence + table_header) having up to a character limit each."
+
+    # Before the table, there should be a blank line
+    intro_sentence = _ensure_blank_line_suffix(intro_sentence)
+
+    intro = intro_sentence + table_header
+    item_max_seq_length = token_limit - len(tokenize(intro))
+    assert item_max_seq_length > 0, item_max_seq_length
+    print("item_max_seq_length", item_max_seq_length)
+    for row in table_rows:
+        print("row", row)
+    chunks = [
+        intro + some_table_rows
+        for some_table_rows in _join_up_to_max_seq_length(
+            table_rows, item_max_seq_length, delimiter=join_delimiter
+        )
+    ]
+    assert all(len(tokenize(chunk)) <= token_limit for chunk in chunks)
+    return chunks
+
+
+def _join_up_to_max_seq_length(
+    lines: Sequence[str], max_seq_length: int, delimiter: str = " "
+) -> list[str]:
+    chunks = []
+    chunk = ""
+    for line in lines:
+        test_chunk = delimiter.join([chunk, line]) if chunk else line
+        token_count = len(tokenize(test_chunk))
+        logger.debug("%i / %i : %r", token_count, max_seq_length, test_chunk)
+        if token_count > max_seq_length:
+            # Don't use test_chunk; add current chunk; start a new chunk
+            if chunk:
+                chunks.append(chunk)
+
+            if (line_token_count := len(tokenize(line))) <= max_seq_length:
+                chunk = line
+            else:
+                logger.warning("Line is too long (%i tokens): %s", line_token_count, line)
+                chunk = ""
+        else:
+            chunk = test_chunk
+
+    # Add the last chunk
+    chunks.append(chunk)
+
+    assert all(len(tokenize(chunk)) <= max_seq_length for chunk in chunks)
+    return chunks
